Rename observational to constrained
