+++ chirho/constrained/internals.py
@@ -4,7 +4,7 @@ import pyro
 import pyro.distributions
 import torch
 
-from chirho.observational.ops import AtomicObservation, observe
+from chirho.constrained.ops import AtomicObservation, observe
 
 T = TypeVar("T")
