--- chirho/observational/__init__.py
@@ -1,1 +1,0 @@
-import chirho.observational.internals  # noqa: F401
