+++ docs/source/constrained.rst
@@ -1,0 +1,13 @@
+Constrained
+===========
+
+.. automodule:: chirho.constrained
+   :members:
+   :undoc-members:
+
+Handlers
+--------
+
+.. automodule:: chirho.constrained.handlers
+   :members:
+   :undoc-members:
