--- docs/source/observational.rst
@@ -1,13 +1,0 @@
-Observational
-=============
-
-.. automodule:: chirho.observational
-   :members:
-   :undoc-members:
-
-Handlers
---------
-
-.. automodule:: chirho.observational.handlers
-   :members:
-   :undoc-members:
